Process request (#24)
* refactor 3

* flag control

* Process request

* FIFO | Read and Write

+++ Api/namedPipe/namedPipe.h
@@ -19,7 +19,7 @@
 
 //the client should use this function to start a request
 //request is initialized and sent to the server
-Request * createRequest(int action, int fd, size_t dataSize, void * data);
+Request * createRequest(int action, int fd, int dataSize, void * data);
 
 Connection * createConnection(int fd);
